move applog to pkg
