+++ go.mod
@@ -6,7 +6,7 @@ require (
 	cloud.google.com/go v0.56.0
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
 	github.com/gorilla/mux v1.7.4
-	github.com/labstack/echo v3.3.10+incompatible
+	github.com/labstack/echo v3.3.10+incompatible // indirect
 	github.com/labstack/gommon v0.3.0 // indirect
 	github.com/spf13/viper v1.6.2
 	go.opencensus.io v0.22.3
